terraform/python_scripts/requirements.txt
bcrypt==4.2.1
boto3==1.33.13
PyYAML==6.0.1
